+++ fuse_viz/include/fuse_viz/serialized_graph_display.h
@@ -1,7 +1,7 @@
 /*
  * Software License Agreement (BSD License)
  *
- *  Copyright (c) 2019, Clearpath Robotics
+ *  Copyright (c) 2022, Locus Robotics
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -32,94 +32,11 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef FUSE_VIZ_SERIALIZED_GRAPH_DISPLAY_H
-#define FUSE_VIZ_SERIALIZED_GRAPH_DISPLAY_H
+#ifndef FUSE_VIZ__SERIALIZED_GRAPH_DISPLAY_H_
+#define FUSE_VIZ__SERIALIZED_GRAPH_DISPLAY_H_
 
-#ifndef Q_MOC_RUN
-#include <fuse_core/graph_deserializer.hpp>
-#include <fuse_core/uuid.hpp>
-#include <fuse_msgs/msg/serialized_graph.hpp>
+#warning This header is obsolete, please include fuse_viz/serialized_graph_display.hpp instead
 
-#include <rviz_common/message_filter_display.hpp>
+#include <fuse_viz/serialized_graph_display.hpp>
 
-#include <OgreColourValue.h>
-#include <OgreSceneNode.h>
-#endif  // Q_MOC_RUN
-
-#include <map>
-#include <memory>
-#include <string>
-#include <unordered_map>
-
-namespace fuse_viz
-{
-
-class Pose2DStampedVisual;
-class RelativePose2DStampedConstraintVisual;
-
-using rviz_common::properties::BoolProperty;
-
-class Pose2DStampedProperty;
-class RelativePose2DStampedConstraintProperty;
-
-/**
- * @brief An rviz dispaly for fuse_msgs::msg::SerializedGraph messages.
- */
-class SerializedGraphDisplay : public rviz_common::MessageFilterDisplay<fuse_msgs::msg::SerializedGraph>
-{
-  Q_OBJECT
-public:
-  SerializedGraphDisplay();
-
-  ~SerializedGraphDisplay() override;
-
-  void reset() override;
-
-protected:
-  void onInitialize() override;
-
-  void onEnable() override;
-
-  void onDisable() override;
-
-  void load(const rviz_common::Config& config) override;
-
-private Q_SLOTS:
-  void updateShowVariables();
-  void updateShowConstraints();
-
-private:
-  using ChangedByUUIDMap = std::unordered_map<fuse_core::UUID, bool, fuse_core::uuid::hash>;
-  using ConstraintByUUIDMap =
-      std::unordered_map<fuse_core::UUID, std::shared_ptr<RelativePose2DStampedConstraintVisual>,
-                         fuse_core::uuid::hash>;
-  using ColorBySourceMap = std::unordered_map<std::string, Ogre::ColourValue>;
-  using ConstraintPropertyBySourceMap = std::map<std::string, RelativePose2DStampedConstraintProperty*>;
-  using ConfigBySourceMap = std::unordered_map<std::string, rviz_common::Config>;
-
-  void clear();
-
-  void processMessage(fuse_msgs::msg::SerializedGraph::ConstSharedPtr msg) override;
-
-  Ogre::SceneNode* root_node_;
-
-  ConstraintByUUIDMap constraint_visuals_;
-
-  ColorBySourceMap source_color_map_;
-
-  ChangedByUUIDMap variables_changed_map_;
-  ChangedByUUIDMap constraints_changed_map_;
-
-  BoolProperty* show_variables_property_;
-  BoolProperty* show_constraints_property_;
-  Pose2DStampedProperty* variable_property_;
-  ConstraintPropertyBySourceMap constraint_source_properties_;
-
-  ConfigBySourceMap constraint_source_configs_;
-
-  fuse_core::GraphDeserializer graph_deserializer_;
-};
-
-}  // namespace fuse_viz
-
-#endif  // FUSE_VIZ_SERIALIZED_GRAPH_DISPLAY_H
+#endif  // FUSE_VIZ__SERIALIZED_GRAPH_DISPLAY_H_
