+++ fuse_viz/src/pose_2d_stamped_property.cpp
@@ -32,8 +32,8 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <fuse_viz/pose_2d_stamped_property.h>
-#include <fuse_viz/pose_2d_stamped_visual.h>
+#include <fuse_viz/pose_2d_stamped_property.hpp>
+#include <fuse_viz/pose_2d_stamped_visual.hpp>
 
 #include <fuse_constraints/relative_pose_2d_stamped_constraint.hpp>
 #include <fuse_core/uuid.hpp>
